Abstracted out beam-search-like methods
